New NPMS Paths form based on the old Wildflower Count form.
Main alteration is support for Surveys 1 and 2.

Note I did change the old form to hard code in a language 'eng' as I couldn't see another way round this as the system was complaining. Also changed the functions from private to protected, but otherwise the original form is largely unchanged. The original ersion can be access through history anyway as I know the client wants it archived.

There are a couple of bugs outstanding I will be dealing with, but am committing as the files are now in a clean state.

1. There is an error on the page that I don't think is affecting functionality.
2. The radio buttons on the Your Paths page are being lost on survey 1 after survey 2 is saved.