Initialised a Drupal iform mobile module.
With the mobile application development expanding, for the best project management and overall clarity it is the best to keep the related files together. 
This module files were part of iform module (iform/media/js/mobile), but since in a short time it has grown from two files to a bunch of libraries, and because I feel it is the best to keep the iform module as slim, but extensible as possible it has found its new home.

Iform mobile module, will be responsible for providing the necessary support for mobile-app related functionality (GPS, storage, communications etc.) and integration with iform and mobile_auth modules (recording and authentication).

It will consist of both JavaScript and PHP libraries integrated and supporting each other.